Refactor and document ComparatorMode class
