+++ qlbm/components/spacetime/initial/volumetric.py
@@ -9,9 +9,9 @@ from qiskit.circuit.library import MCMTGate, XGate
 from typing_extensions import override
 
 from qlbm.components.base import LBMPrimitive
-from qlbm.components.ms.primitives import Comparator, ComparatorMode
+from qlbm.components.ms.primitives import Comparator
 from qlbm.lattice.lattices.spacetime_lattice import SpaceTimeLattice
-from qlbm.tools.utils import flatten
+from qlbm.tools.utils import ComparatorMode, flatten
 
 
 class VolumetricSpaceTimeInitialConditions(LBMPrimitive):
